tests: add mysql prebuilt tests to cloud sql mysql integration
